Fix double execution of console commands during deployment
This was due to repeating "on release_roles" in both the task and the
function. A third parameter has been introduced to symfony_console to
maintain compatibility for executing Symfony commands which don't have
their own task.

Additionally, "within release_path" is not required since it's already
part of the absolute path passed to execute.